All game information streamlined without refresh
Implemented websockets gem. All in-game data is captured with Javascript
and sent through an EventMapper to custom Rails Websockets controllers.

There is a skeleton combat system in models/round.rb in which to build
this new interactive system around. Right now it incorporates
multithreading, this is just until we can get the real timing system up
and running in the next branch.