Users can play via either HTTP or Websockets
The game is now playable by either using HTTP or websockets. Although
the only practical use case for HTTP is when a user initially logins and
then is routed through the dashboards_controller, or when a user
refreshes the page midgame. The websockets kick in when the user moves his character for the first
and each time thereafter.

Multithreading has officially been removed. Timing will be implemented
in the next branch without multithreading.

Javascript has been refactored into multiples files. The files are
organized based on their purpose. Rather than loading each one when the
page loads, all the files are wrapped in a function, and all those
functions are placed in a mater javascript file named Pusher. Pusher is
called independently in the Javascript pipeline tree. Pusher.js contains
the functionality that websockets uses. The other primary standalone
javascript file is websocketsHelpers.js. This file is also in the
Javascript pipeline tree. A method qualifies to be apart of
websocketsHelpers.js if there are functions in Pusher.js which depend on
it, but it is not totally relevant to websockets in of itself, or if
multiple functions within websockets rely on it.

Our Javascript needs to handle our data that is received from the
server. Since we are using websockets, rendering partials in ERB is no
longer an option. The way I am handling dynamic content generation in
Javascript, where I would normally be able to use a partial, is to
instead create a template of raw html and insert the data received into
it. To push the generated HTML into the view we can either append it or
set the elements html property to the generated HTML. Whether we append
the data or set the html property is obviously dependent on the context
of the element in question.