Movement is disabled during attack
Note: It is essential to keep in mind when working with Javascripts
setTimeout and setInterval functions that these do not behave like
sleep, in that normal execution control is not paused or interrupted in
any way. SetTimeout and setInterval work like psuedo-multithreads, which
go off on their own execution timeline.